Fix strange label select parent behavior
The problem is that the node label can not have the type "node:label", because all types with starting "node" will automatically have some features added. This results that the label had the moveFeature etc. and this one was ecential for the desicion if the resize handles should be shown on the parent.
Quick fix: rename the node label type to "node-label"