Remove unnecessary @injectable annotation and yarn resolutions
The emitted JavaScript code has proper meta-data but accesses the a
class before it was declared/initialized since it is used in the
constructor arguments. We can re-order the declarations or remove the
injectable annotation to avoid the wrong access. We do both.

Remove unnecessary yarn resolutions and update yarn.lock