XIVY-13010 Fix regresion bug on direct move an element
The fix about do not reevaluate the resize element (e.g. when the mouse leaves the window) broke the ability to direct move an element. If you didn't select an element first, but directly draged it to a new location, this update wasn't sent to the backend.
Now we add this lane fix only, the activeResizeElement is an instance of a LaneNode